+++ CHANGELOG.md
@@ -9,6 +9,11 @@ All notable changes to MiniJinja are documented here.
 - Add missing escape support for single quotes (`'`).  (#81) 
 - Added support for newlines in string literals.  (#85)
 - Added support for block assignment syntax.  (#86)
+- Added string concatenatino with `+` for Jinja compat.  (#87)
+- Enable debug mode by default in debug builds.  (#88)
+- Added `render!` macro and `render_str`.  (#89)
+- Fixed an issue where trailing whitespace removal did not work on blocks.  (#90)
+- Added `loop.changed()` method.  (#91)
 
 # 0.18.1